+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+### Fixes
+* Atmospheric morado leaves not timbering
+
 ## 4.17.9
 ### Changes
 * Sweeping overhaul